Added G28 bug
As writing a post on the forum about the current broken endstop implementation, I figured out how to do it *right* so it works consistent and as expected in all circumstances. Requires a new M-code...

Thought I'd record this so I don't forget to write it in if nobody else does first.